+++ packages/frontend/pages/lp.tsx
@@ -1,214 +1,117 @@
-import { createStyles, makeStyles } from '@material-ui/core'
-import Typography from '@material-ui/core/Typography'
-import Image from 'next/image'
-import React, { useState } from 'react'
-
-import squeethTokenSymbol from '../public/images/Squeeth.svg'
-import ObtainSqueeth from '@components/Lp/ObtainSqueeth'
-import SqueethInfo from '@components/Lp/SqueethInfo'
-import LPBuyChart from '@components/Charts/LPBuyChart'
-import LPMintChart from '@components/Charts/LPMintChart'
+import LPInfoCard from '@components/Lp/LPInfoCard'
 import Nav from '@components/Nav'
-import { LPProvider } from '@context/lp'
-import { useRestrictUser } from '@context/restrict-user'
-import { SqueethTab, SqueethTabs } from '@components/Tabs'
-import { useETHPrice } from '@hooks/useETHPrice'
+import { Typography, Box, Grid } from '@material-ui/core'
+import { createStyles, makeStyles } from '@material-ui/core/styles'
+import React from 'react'
 
 const useStyles = makeStyles((theme) =>
   createStyles({
     container: {
-      padding: theme.spacing(4, 10),
-      marginLeft: 'auto',
-      marginRight: 'auto',
-      maxWidth: '1500px',
-      display: 'flex',
-      justifyContent: 'space-between',
+      padding: theme.spacing(4, 12),
     },
-    logoContainer: {
-      display: 'flex',
-      alignItems: 'center',
-      width: '100%',
-    },
-    logoTitle: {
-      marginLeft: theme.spacing(1),
-      [theme.breakpoints.down('sm')]: {
-        fontSize: 18,
-      },
-    },
-    logoSubTitle: {
-      marginLeft: theme.spacing(1),
-      [theme.breakpoints.down('sm')]: {
-        fontSize: 16,
-      },
-    },
-    logo: {
-      marginTop: theme.spacing(0.5),
-      alignSelf: 'flex-start',
-    },
-    comparison: {
-      marginTop: theme.spacing(4),
-      //paddingBottom: theme.spacing(2),
-    },
-    comparisonItem: {
-      width: '400px',
-    },
-    comparisonPoint: {
-      marginTop: theme.spacing(2),
-      padding: theme.spacing(0, 2),
-    },
-    details: {
-      marginTop: theme.spacing(4),
-      // paddingLeft: theme.spacing(4),
-      // display: 'flex',
+    subHeading: {
+      fontSize: '15px',
+      color: theme.palette.text.secondary,
+      lineHeight: '140%',
     },
     heading: {
-      marginTop: theme.spacing(3),
+      fontSize: '34px',
+      lineHeight: '140%',
+      fontWeight: 700,
     },
-    tradeForm: {
-      position: 'sticky',
-      top: '100px',
+    poolValue: {
+      fontSize: '27px',
+      lineHeight: '140%',
+      fontWeight: 700,
     },
-    chartNav: {
-      border: `1px solid ${theme.palette.primary.main}30`,
+    tradeCard: {
+      backgroundColor: theme.palette.background.lightStone,
+      padding: theme.spacing(2),
+      borderRadius: theme.spacing(2),
+      height: '100%',
+    },
+    infoBox: {
+      padding: theme.spacing(2, 3),
+      borderRadius: theme.spacing(2),
+      border: `1px solid ${theme.palette.background.border}`,
+    },
+    poolDataItem: {
+      borderRight: `1px solid ${theme.palette.background.border}`,
+      padding: theme.spacing(0, 3),
+      '&:last-child': {
+        border: 'none',
+      },
+      '&:first-child': {
+        paddingLeft: 0,
+      },
     },
   }),
 )
 
-export function LPCalculator() {
+const LP: React.FC = () => {
   const classes = useStyles()
-  const { isRestricted } = useRestrictUser()
-  const ethPrice = useETHPrice()
-  const [lpType, setLpType] = useState(0)
 
   return (
     <div>
       <Nav />
       <div className={classes.container}>
-        <div style={{ width: '800px', paddingRight: '16px' }}>
-          <div style={{ display: 'flex' }}>
-            <div className={classes.logo}>
-              <Image src={squeethTokenSymbol} alt="squeeth token" width={37} height={37} />
-            </div>
-            <div>
-              <Typography variant="h5" className={classes.logoTitle}>
-                Uniswap V3 LP SQTH-ETH Pool
-              </Typography>
-              <Typography className={classes.logoSubTitle} variant="body1" color="textSecondary">
-                Earn LP fees for providing SQTH-ETH liquidity
-              </Typography>
-            </div>
-          </div>
-          <SqueethInfo />
-          <div className={classes.details}>
-            <div style={{ display: 'flex' }}>
-              <SqueethTabs
-                style={{ background: 'transparent' }}
-                className={classes.chartNav}
-                value={lpType}
-                onChange={(evt, val) => setLpType(val)}
-                aria-label="Sub nav tabs"
-                scrollButtons="auto"
-                variant="scrollable"
-              >
-                <SqueethTab label={'Buy and LP'} />
-                <SqueethTab label="Mint and LP" />
-              </SqueethTabs>
-            </div>
-            {lpType === 0 ? (
-              <div style={{ marginTop: '16px' }}>
-                <Typography color="primary" variant="h6">
-                  Buy squeeth and LP
-                </Typography>
-                <Typography>
-                  Earn a payoff similar to ETH<sup>1.5</sup>
-                </Typography>
-                <Typography className={classes.heading} variant="subtitle1" color="primary">
-                  Details
-                </Typography>
-                <Typography>
-                  Buying and LPing gives you a leverage position with a payoff similar to ETH<sup>1.5</sup>. You give up
-                  some of your squeeth upside in exchange for trading fees. You are paying funding for being long
-                  squeeth, but earning fees from LPing on Uniswap.
-                </Typography>
-                <Typography className={classes.heading} variant="subtitle1" color="primary">
-                  Payoff
-                </Typography>
-                <LPBuyChart ethPrice={ethPrice.toNumber()} />
-                <Typography variant="caption" color="textSecondary">
-                  This payoff diagram does not included funding or trading fees and assumes implied volatility stays
-                  constant.{' '}
-                </Typography>
-                <Typography className={classes.heading} variant="subtitle1" color="primary">
-                  Risks
-                </Typography>
-                <Typography variant="body1">
-                  You are exposed to squeeth funding, so if you hold the position for a long period of time without
-                  upward price movements in ETH, you can lose considerable funds to funding payments.
-                </Typography>
-                <br />
-                <Typography variant="body1">
-                  {' '}
-                  Squeeth smart contracts have been audited by Trail of Bits, Akira, and Sherlock. However, smart
-                  contracts are experimental technology and we encourage caution only risking funds you can afford to
-                  lose.
-                </Typography>
-              </div>
-            ) : (
-              <div style={{ marginTop: '16px' }}>
-                <Typography color="primary" variant="h6">
-                  Mint squeeth and LP
-                </Typography>
-                <Typography>Earn yield from trading fees while being long ETH</Typography>
-                <Typography className={classes.heading} variant="subtitle1" color="primary">
-                  Details
-                </Typography>
-                <Typography>
-                  Minting and LPing is similar to a covered call. You start off with a position similar to 1x long ETH
-                  that gets less long ETH as the price moves up and longer ETH as the price moves down.
-                </Typography>
-                <Typography className={classes.heading} variant="subtitle1" color="primary">
-                  Payoff
-                </Typography>
-                <LPMintChart ethPrice={ethPrice.toNumber()} />
-                <Typography variant="caption" color="textSecondary">
-                  This payoff diagram does not included funding or trading fees and assumes implied volatility stays
-                  constant.{' '}
-                </Typography>
-                <Typography className={classes.heading} variant="subtitle1" color="primary">
-                  Risks
-                </Typography>
-                <Typography variant="body1">
-                  You enter this position neutral to squeeth exposure, but could end up long squeeth exposed to funding
-                  or short squeeth depending on ETH price movements. If you fall below the minimum collateralization
-                  threshold (150%), you are at risk of liquidation.
-                </Typography>
-                <br />
-                <Typography variant="body1">
-                  {' '}
-                  Squeeth smart contracts have been audited by Trail of Bits, Akira, and Sherlock. However, smart
-                  contracts are experimental technology and we encourage caution only risking funds you can afford to
-                  lose.
-                </Typography>
-              </div>
-            )}
-          </div>
-        </div>
-        <div>
-          <div className={classes.tradeForm}>{!isRestricted ? <ObtainSqueeth /> : null}</div>
-        </div>
+        <Box display="flex" alignItems="center" mb={4}>
+          <Box mr={4}>
+            <Typography className={classes.subHeading}>Earn fees passively</Typography>
+            <Typography className={classes.heading}>Become an LP</Typography>
+          </Box>
+          <Box maxWidth={520} pl={4} borderLeft="1px solid #FFFFFF08">
+            <Typography color="textSecondary">
+              Become a liquidity provider in the SQTH-ETH pool on Uniswap v3. You can choose to Buy or Mint the Squeeth
+              (oSQTH) you deposit.
+            </Typography>
+          </Box>
+        </Box>
+        <Grid container spacing={4}>
+          <Grid item xs={12} lg={6}>
+            <Grid container spacing={2}>
+              <Grid item xs={6}>
+                <LPInfoCard title="Liquidation Price" value="$0.00" disabled />
+              </Grid>
+              <Grid item xs={6}>
+                <LPInfoCard title="Realized P&L" value="$0.00" disabled />
+              </Grid>
+              <Grid item xs={6}>
+                <LPInfoCard title="Ethereum (ETH) Price" value="$ 4200.50" />
+              </Grid>
+              <Grid item xs={6}>
+                <LPInfoCard title="Position Size(oSQTH)" value="0.00" disabled />
+              </Grid>
+              <Grid item xs={12}>
+                <Box className={classes.infoBox}>
+                  <Typography className={classes.subHeading} style={{ marginBottom: '16px' }}>
+                    Pool Details
+                  </Typography>
+                  <Box display="flex">
+                    <div className={classes.poolDataItem}>
+                      <Typography className={classes.poolValue}>$133.4m</Typography>
+                      <Typography className={classes.subHeading}>ETH Deposited</Typography>
+                    </div>
+                    <div className={classes.poolDataItem}>
+                      <Typography className={classes.poolValue}>$33.4m</Typography>
+                      <Typography className={classes.subHeading}>oSQTH Minted</Typography>
+                    </div>
+                    <div className={classes.poolDataItem}>
+                      <Typography className={classes.poolValue}>1.2k</Typography>
+                      <Typography className={classes.subHeading}>Depositors</Typography>
+                    </div>
+                  </Box>
+                </Box>
+              </Grid>
+            </Grid>
+          </Grid>
+          <Grid item xs={12} lg={5}>
+            <Box className={classes.tradeCard}>Trade Page goes here</Box>
+          </Grid>
+        </Grid>
       </div>
     </div>
   )
 }
 
-export function LPage() {
-  return (
-    // <TradeProvider>
-    <LPProvider>
-      <LPCalculator />
-    </LPProvider>
-    // </TradeProvider>
-  )
-}
-
-export default LPage
+export default LP
