+++ packages/frontend/src/components/Lp/LPInfoCard.test.tsx
@@ -1,0 +1,7 @@
+import { render, screen } from '@testing-library/react'
+
+describe('LPInfoCard', () => {
+  it('renders label text', () => {
+    expect(0).toBe(1)
+  })
+})
